Seperated Dupe code into util
